added possibility to run calc_unbalanced for rdm movies
